scripts/requirements.txt
# HOC API Explorer Requirements

httpx>=0.27.0
python-dotenv>=1.0.0

